requirements.txt
Flask==1.1.2
gunicorn==19.9.0
grpcio==1.32.0
six==1.15.0
h5py==2.10.0
jedi==0.17.2
Jinja2==2.11.2
scipy==1.4.1
Keras-Preprocessing==1.1.2
matplotlib==3.4.0
numpy==1.19.5
pandas==1.2.3
pandas-datareader==0.9.0
pandocfilters==1.4.3
pickleshare==0.7.5
Pillow==8.1.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
Werkzeug==1.0.1
